# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trademark Ledger & Case Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List / search clients
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: object
                required: [data, total]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Client" }
                  total: { type: integer }
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClientInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClientUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeletedResponse" }
  /clients/{id}/cases:
    get:
      operationId: getClientCases
      tags: [clients]
      summary: Get all cases for a client
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cases for the client
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TrademarkCase" }
  /clients/{id}/ledger:
    get:
      operationId: getClientLedger
      tags: [clients]
      summary: Get consolidated ledger for a client (all cases)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Consolidated ledger
          content:
            application/json:
              schema:
                type: object
                required: [entries, totalDue, totalReceived, balance]
                properties:
                  entries:
                    type: array
                    items: { $ref: "#/components/schemas/LedgerEntry" }
                  totalDue: { type: number }
                  totalReceived: { type: number }
                  balance: { type: number }

  # ── TRADEMARK CASES ──────────────────────────────────────────────────────
...
  /import/cases:
    post:
      operationId: importCases
      tags: [import-export]
      summary: Bulk import trademark cases from Excel data (JSON rows)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BulkImportInput" }
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportResult" }
  /export/cases:
    get:
      operationId: exportCases
      tags: [import-export]
      summary: Export cases as JSON for Excel generation
      parameters:
        - name: clientId
          in: query
          schema: { type: integer }
        - name: stage
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Cases export data
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TrademarkCase" }
  /export/ledger:
    get:
      operationId: exportLedger
      tags: [import-export]
      summary: Export ledger entries as JSON for Excel generation
      parameters:
        - name: clientId
          in: query
          schema: { type: integer }
        - name: folderId
          in: query
          schema: { type: string }
        - name: startDate
          in: query
          schema: { type: string, format: date }
        - name: endDate
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Ledger export data
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LedgerEntry" }

  # ── SEARCH ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    DeletedResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }

    # ── CLIENT ──────────────────────────────────────────────────────────────
    Client:
      type: object
      required: [id, clientNumber, name, createdAt, updatedAt]
      properties:
        id: { type: integer }
        clientNumber: { type: string }
        name: { type: string }
        address: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ClientInput:
      type: object
      required: [clientNumber, name]
      properties:
        clientNumber: { type: string }
        name: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
        notes: { type: string }
    ClientUpdate:
      type: object
      properties:
        clientNumber: { type: string }
        name: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
        notes: { type: string }

    # ── TRADEMARK CASE ───────────────────────────────────────────────────────
    TrademarkCase:
      type: object
      required: [id, folderNumber, clientId, applicantName, stage, createdAt, updatedAt]
      properties:
        id: { type: integer }
        folderNumber: { type: string }
        clientId: { type: integer }
        clientNumber: { type: ["string", "null"] }
        clientName: { type: ["string", "null"] }
        tmNumber: { type: ["string", "null"] }
        applicantName: { type: string }
        class: { type: ["string", "null"] }
        filingDate: { type: ["string", "null"] }
        stage: { type: integer }
        subStage: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    LedgerEntry:
      type: object
      required: [id, caseId, date, folderNumber, detail, dueAmount, receivedAmount, runningBalance, createdAt]
      properties:
        id: { type: integer }
        caseId: { type: integer }
        date: { type: string }
        folderNumber: { type: string }
        tmNumber: { type: ["string", "null"] }
        stage: { type: ["integer", "null"] }
        detail: { type: string }
        dueAmount: { type: number }
        receivedAmount: { type: number }
        runningBalance: { type: number }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    BulkImportInput:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            type: object
            properties:
              folderNumber: { type: string }
              clientNumber: { type: string }
              tmNumber: { type: string }
              applicantName: { type: string }
              class: { type: string }
              filingDate: { type: string }
              stage: { type: integer }
              subStage: { type: string }
              notes: { type: string }
    ImportResult:
      type: object
      required: [imported, skipped, errors]
      properties:
        imported: { type: integer }
        skipped: { type: integer }
        errors:
          type: array
          items:
            type: object
            properties:
              row: { type: integer }
              error: { type: string }

    # ── SEARCH ───────────────────────────────────────────────────────────────
...